% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_scraping_helpers.R
\name{cn_conditional_article_scrape}
\alias{cn_conditional_article_scrape}
\title{Conditionally read COVID-19 Daily Bulletin
Used by clusterMap and other functions to read in
only those articles that haven't been read in yet.}
\usage{
cn_conditional_article_scrape(href, content)
}
\description{
Conditionally read COVID-19 Daily Bulletin
Used by clusterMap and other functions to read in
only those articles that haven't been read in yet.
}
