% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scraping_helpers.R
\name{cn_page_link_scrape}
\alias{cn_page_link_scrape}
\title{Scrape links to NHC COVID-19 Daily Bulletin}
\usage{
cn_page_link_scrape(html_page, root_url = "http://en.nhc.gov.cn/")
}
\description{
Scrape links to NHC COVID-19 Daily Bulletin
}
