% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scraping_helpers.R
\name{href_to_date}
\alias{href_to_date}
\title{Helper to convert URL to Date based on assumed format
Yeah yeah, I could have used str_match, but then I would have to install stringi
each time on CI/CD. That's a PITA.}
\usage{
href_to_date(x)
}
\description{
Helper to convert URL to Date based on assumed format
Yeah yeah, I could have used str_match, but then I would have to install stringi
each time on CI/CD. That's a PITA.
}
