% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scraping_helpers.R
\name{read_page}
\alias{read_page}
\title{A simple HTML page reader with debouncer
This is a stupidly simple function to get around
rate-limit or anti-crawling measures by dynamically
waiting between each request for a set time.

It's heuristic, and may still return a try-error that
must be handled downstream. Ideally I'd use OOP and handle
this at the app level instead of the request level.}
\usage{
read_page(page_url, max_wait = 20)
}
\arguments{
\item{page_url}{(character) A URL we want to read}

\item{max_wait}{(numeric) Maximum debounce}
}
\description{
A simple HTML page reader with debouncer
This is a stupidly simple function to get around
rate-limit or anti-crawling measures by dynamically
waiting between each request for a set time.

It's heuristic, and may still return a try-error that
must be handled downstream. Ideally I'd use OOP and handle
this at the app level instead of the request level.
}
\note{
max_wait
}
